Retry setsocketopt calls with SO_SNDBUF/SO_RCVBUF
criu needs to be invoked with --unprivileged in an
unprivileged container environment. This patch proposes
a simple work-around to have the setsockopt() calls
retried with SO_SNDBUF/SO_RCVBUF when the calls
with SO_SNDBUFFORCE/SO_RCVBUFFORCE fail with EPERM, which
means we are very likely running in a container.